+++ CHANGELOG.md
@@ -5,6 +5,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [next-version] - 2020-11-xx
+
+### Added
+
+- TwoDimensionPosition struct to represent a position in 2D spaces, focused on positions within arrays.
+
+### Changed
+
+- System.Drawing.Point has been replaced by TwoDimensionPosition struct.
+
+### Removed
+
+- All uses of System.Drawing.Point struct.
+
+### Fixed
+
+- Misunderstandings about the x and y axes in an array.
+
 ## [0.1.0] - 2020-11-04
 
 ### Added